Refactor backup create
+ add useTypedRoute hook
