+++ src/screens/backup-create.tsx
@@ -1,142 +1,18 @@
-import React, {useState} from 'react';
+import React from 'react';
 
-import {RouteProp, useNavigation, useRoute} from '@react-navigation/native';
-import {StackNavigationProp} from '@react-navigation/stack';
-import {StyleSheet, View} from 'react-native';
-
-import {MnemonicWord} from '@app/components/mnemonic-word';
-import {
-  Button,
-  ButtonVariant,
-  Checkbox,
-  Copy,
-  CopyButton,
-  InfoBlock,
-  InfoBlockType,
-  PopupContainer,
-  Spacer,
-  Text,
-} from '@app/components/ui';
-import {HapticEffects, vibrate} from '@app/services/haptic';
-import {RootStackParamList} from '@app/types';
-import {
-  LIGHT_BG_3,
-  LIGHT_TEXT_BASE_2,
-  LIGHT_TEXT_GREEN_1,
-} from '@app/variables';
+import {BackupCreate} from '@app/components/backup-create';
+import {useTypedNavigation, useTypedRoute} from '@app/hooks';
 
 export const BackupCreateScreen = () => {
-  const navigation = useNavigation<StackNavigationProp<RootStackParamList>>();
-  const route = useRoute<RouteProp<RootStackParamList, 'backupCreate'>>();
+  const navigation = useTypedNavigation();
+  const {mnemonic, rootAddress} = useTypedRoute<'backupCreate'>().params;
 
-  const [checked, setChecked] = useState(false);
-
-  const onClickCheck = (val: boolean) => {
-    vibrate(HapticEffects.impactLight);
-    setChecked(val);
+  const onSubmit = () => {
+    navigation.navigate('backupVerify', {
+      rootAddress,
+      mnemonic,
+    });
   };
 
-  return (
-    <PopupContainer style={page.container}>
-      <Text t4 style={page.t4}>
-        Your recovery phrase
-      </Text>
-      <Text t11 color={LIGHT_TEXT_BASE_2} center>
-        Write down or copy these words in the right order and save them
-        somewhere safe.
-      </Text>
-      <Spacer style={page.space}>
-        <View style={page.mnemonics}>
-          <View style={page.column}>
-            {route.params.mnemonic
-              .split(' ')
-              .slice(0, 6)
-              .map((t, i) => (
-                <MnemonicWord key={`${t}${i}`} word={t} index={i + 1} />
-              ))}
-          </View>
-          <View style={page.column}>
-            {route.params.mnemonic
-              .split(' ')
-              .slice(6, 12)
-              .map((t, i) => (
-                <MnemonicWord key={`${t}${i}`} word={t} index={i + 7} />
-              ))}
-          </View>
-        </View>
-        <CopyButton value={route.params.mnemonic ?? ''} style={page.copy}>
-          <Copy color={LIGHT_TEXT_GREEN_1} />
-          <Text clean style={page.copyText}>
-            Copy
-          </Text>
-        </CopyButton>
-      </Spacer>
-      <InfoBlock t15 type={InfoBlockType.warning} style={page.marginBottom}>
-        If you lose your recovery phrase, you will be unable to access your
-        funds, as nobody will be able to restore it.
-      </InfoBlock>
-      <View style={page.agree}>
-        <Checkbox value={checked} onPress={onClickCheck}>
-          <Text t14 style={page.agreeText}>
-            I understand that if I lose my recovery phrase, I will not be able
-            to restore access to my account
-          </Text>
-        </Checkbox>
-      </View>
-      <Button
-        title="Continue"
-        style={page.submit}
-        variant={ButtonVariant.contained}
-        disabled={!checked}
-        onPress={() =>
-          navigation.navigate('backupVerify', {
-            rootAddress: route.params.rootAddress,
-            mnemonic: route.params.mnemonic,
-          })
-        }
-      />
-    </PopupContainer>
-  );
+  return <BackupCreate onSubmit={onSubmit} />;
 };
-
-const page = StyleSheet.create({
-  container: {
-    paddingHorizontal: 20,
-    paddingTop: 20,
-  },
-  mnemonics: {
-    backgroundColor: LIGHT_BG_3,
-    flexDirection: 'row',
-    padding: 20,
-    borderRadius: 16,
-  },
-  column: {flex: 1},
-  marginBottom: {marginBottom: 20},
-  space: {justifyContent: 'center'},
-  agree: {marginBottom: 4, flexDirection: 'row'},
-  agreeText: {
-    flex: 1,
-    color: LIGHT_TEXT_BASE_2,
-    fontWeight: '600',
-    marginLeft: 12,
-    marginBottom: 4,
-  },
-  submit: {marginVertical: 16},
-  copy: {
-    paddingVertical: 12,
-    paddingHorizontal: 32,
-    marginHorizontal: 4,
-    alignSelf: 'center',
-  },
-  copyText: {
-    color: LIGHT_TEXT_GREEN_1,
-    fontWeight: '700',
-    fontSize: 16,
-    lineHeight: 22,
-    marginHorizontal: 4,
-  },
-  t4: {
-    alignSelf: 'center',
-    alignItems: 'center',
-  },
-});
